+++ client/test/node/components/Pdf-test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { expect } from 'chai';
 import { mount } from 'enzyme';
-import { Pdf, getInitialAnnotationIconPageCoords } from '../../../app/components/Pdf';
+import { Pdf } from '../../../app/components/Pdf';
 import sinon from 'sinon';
 import _ from 'lodash';
 
@@ -84,92 +84,6 @@ describe('Pdf', () => {
         it('pages are not redrawn (no-op)', () => {
           wrapper.setProps({ id: 'newId' });
           expect(draw.callCount).to.equal(0);
-        });
-      });
-    });
-  });
-
-  describe('getInitialAnnotationIconPageCoords', () => {
-    describe('zoom = 1', () => {
-      it('centers the icon when the page is contained entirely by the scroll window', () => {
-        const pageBox = {
-          top: 100,
-          bottom: 500,
-          left: 200,
-          right: 300
-        };
-        const scrollWindowBox = {
-          top: 0,
-          bottom: 1000,
-          left: 0,
-          right: 900
-        };
-
-        expect(getInitialAnnotationIconPageCoords(pageBox, scrollWindowBox, 1)).to.deep.equal({
-          y: 180,
-          x: 30
-        });
-      });
-
-      it('centers the icon when the scroll window is contained entirely by the page', () => {
-        const pageBox = {
-          top: -300,
-          bottom: 1000,
-          left: -500,
-          right: 1200
-        };
-        const scrollWindowBox = {
-          top: 0,
-          bottom: 900,
-          left: 0,
-          right: 700
-        };
-
-        expect(getInitialAnnotationIconPageCoords(pageBox, scrollWindowBox, 1)).to.deep.equal({
-          y: 730,
-          x: 830
-        });
-      });
-    });
-
-    describe('zoom = 2', () => {
-      it('centers the icon when the page is contained entirely by the scroll window', () => {
-        const pageBox = {
-          top: 100,
-          bottom: 500,
-          left: 200,
-          right: 300
-        };
-        const scrollWindowBox = {
-          top: 0,
-          bottom: 1000,
-          left: 0,
-          right: 900
-        };
-
-        expect(getInitialAnnotationIconPageCoords(pageBox, scrollWindowBox, 2)).to.deep.equal({
-          y: 80,
-          x: 5
-        });
-      });
-
-      it('centers the icon when the scroll window is contained entirely by the page', () => {
-        const pageBox = {
-          top: -300,
-          bottom: 1000,
-          left: -500,
-          right: 1200
-        };
-        const scrollWindowBox = {
-          top: 100,
-          bottom: 900,
-          left: 100,
-          right: 700
-        };
-
-        expect(getInitialAnnotationIconPageCoords(pageBox, scrollWindowBox, 2)).to.deep.equal({
-          y: 380,
-          x: 430
         });
       });
     });
